UI갱신과 리스트뷰 처리에 있어 스레드 Safe하게 수정
클라이언트 강제종료 시 UI를 바로 갱신하도록 함 (운나쁘면 5초동안 강종된 프로세스가 목록에 떠있음)

+++ CPU_Preference_Changer/BackgroundTask/MabiClientKillTask.cs
@@ -1,4 +1,6 @@
-﻿using CPU_Preference_Changer.Core.BackgroundFreqTaskManager;
+﻿using CPU_Preference_Changer.Core;
+using CPU_Preference_Changer.Core.BackgroundFreqTaskManager;
+using CPU_Preference_Changer.Core.SingleTonTemplate;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
@@ -6,7 +8,22 @@ using System.Linq;
 using System.Text;
 
 namespace CPU_Preference_Changer.BackgroundTask {
+    /// <summary>
+    /// 마비노기 클리어언트 종료 Task.
+    /// 지정된 시간이 되면 특정 PID값을 가진 프로세스를 강제종료 함.
+    /// </summary>
     class MabiClientKillTask : IBackgroundFreqTask{
+
+        /// <summary>
+        /// 클라이언트 강종 되었을 때 일어날 이벤트 타입정의
+        /// </summary>
+        /// <param name="sender"></param>
+        public delegate void OnClientKilled(object sender);
+
+        /// <summary>
+        /// 클라이언트를 강종 시켰을 때 일어날 이벤트
+        /// </summary>
+        public event OnClientKilled onClientProcessKilled;
 
         /// <summary>
         /// 예약 종료 지정된 프로세스 PID값 보관
@@ -19,13 +36,22 @@ namespace CPU_Preference_Changer.BackgroundTask {
         private DateTime killTime;
 
         private const ulong freq = 1000; /*1초단위로 여유롭게 감시한다.*/
-
+        
+        /// <summary>
+        /// 클라이언트 종료 Task생성자
+        /// </summary>
+        /// <param name="PID">종효 할 PID</param>
+        /// <param name="killTime">종료 시간</param>
         public MabiClientKillTask(int PID, DateTime killTime)
         {
             this.PID = PID;
             this.killTime = killTime;
         }
 
+        /// <summary>
+        /// 이 Task가 어느정도의 주기로 반복 실행하면 되는지 반환
+        /// </summary>
+        /// <returns></returns>
         public ulong getFreqTick()
         {
             return freq;
@@ -37,9 +63,8 @@ namespace CPU_Preference_Changer.BackgroundTask {
         /// </summary>
         private void killCLientProcess2()
         { 
-
+            /*나중에 잘 안되는 일 생기면 그때 구현.*/
         }
-
 
         /// <summary>
         /// by LT골든힐트
@@ -50,9 +75,11 @@ namespace CPU_Preference_Changer.BackgroundTask {
             try {
                 using (Process p = Process.GetProcessById(PID)) {
                     try {
-                        // kill process
-                        if (p != null)
+                        /*혹시라도 그짧은 순간에 마비가 종료되고 다른 프로세스로 켜졌을 수 있으니 확인해보고 종료*/
+                        if ( (p!=null) && MabiProcess.isMabiProcess(p)) {
+                            // kill process
                             p.Kill();
+                        }
                     } catch {
                         // kill exception or access exception
                         p.Dispose();
@@ -64,6 +91,12 @@ namespace CPU_Preference_Changer.BackgroundTask {
             }
         }
 
+        /// <summary>
+        /// 클라이언트 종료 시간이 되었는지 주기적으로 확인하는 함수
+        /// </summary>
+        /// <param name="taskHandle"></param>
+        /// <param name="param"></param>
+        /// <returns></returns>
         public bool runFreqWork(HBFT taskHandle, object param)
         {
             DateTime curTime = DateTime.Now;
@@ -73,6 +106,9 @@ namespace CPU_Preference_Changer.BackgroundTask {
                 /*예약 시간을 넘었다! 해당 PID값 확인해보고 
                  * 여전히 존재한다면 종료!*/
                 killClientProcess();
+                if (onClientProcessKilled != null) {
+                    onClientProcessKilled(this);
+                }
                 return false;
             }
             return true;
